Castor3D: Reworked GUI controls flags as uint64_t.

+++ include/Core/Castor3D/Gui/Controls/CtrlComboBox.hpp
@@ -45,7 +45,7 @@ namespace castor3d
 			, int selected
 			, castor::Position const & position
 			, castor::Size const & size
-			, uint32_t flags = 0
+			, uint64_t flags = 0
 			, bool visible = true );
 
 		/** Constructor
@@ -67,7 +67,7 @@ namespace castor3d
 			, int selected
 			, castor::Position const & position
 			, castor::Size const & size
-			, uint32_t flags = 0
+			, uint64_t flags = 0
 			, bool visible = true )
 			: Control{ ControlType::eComboBox
 				, scene
